renamed references to wallet -> bitcoin
