+++ lib/flows/bitcoin/send/choose_recipient.dart
@@ -7,7 +7,7 @@ import 'package:orange/components/header.dart';
 import 'package:orange/components/list_item.dart';
 import 'package:orange/components/text_input.dart';
 
-import 'package:orange/flows/wallet/send/amount.dart';
+import 'package:orange/flows/bitcoin/send/amount.dart';
 
 import 'package:orange/util.dart';
 import 'package:orange/classes.dart';
