+++ lib/flows/bitcoin/send/confirm_send.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:orange/theme/stylesheet.dart';
 
-import 'package:orange/flows/wallet/send/confirmation.dart';
-import 'package:orange/flows/wallet/send/amount.dart';
-import 'package:orange/flows/wallet/send/transaction_speed.dart';
-import 'package:orange/flows/wallet/send/send.dart';
+import 'package:orange/flows/bitcoin/send/confirmation.dart';
+import 'package:orange/flows/bitcoin/send/amount.dart';
+import 'package:orange/flows/bitcoin/send/transaction_speed.dart';
+import 'package:orange/flows/bitcoin/send/send.dart';
 
 import 'package:orange/components/interface.dart';
