+++ lib/flows/bitcoin/send/transaction_speed.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:orange/components/interface.dart';
 import 'package:orange/components/radio_selectors.dart';
-import 'package:orange/flows/wallet/send/confirm_send.dart';
+import 'package:orange/flows/bitcoin/send/confirm_send.dart';
 
 import 'package:orange/components/content.dart';
 import 'package:orange/components/header.dart';
