+++ lib/flows/bitcoin/transaction_details.dart
