+++ lib/theme/stylesheet.dart
@@ -161,7 +161,7 @@ class AppPadding {
   static const double bumper = 16;
   static const double valueDisplaySep = 8;
   static const double valueDisplay = 32;
-  static const double navBar = 64;
+  static const double navBar = 32;
   static const double listItem = 16;
   static const double textInput = 12;
   static const double tips = 8;
